Build with -Wextra and -Werror on both CPUs
The ARM7 was already clean under -Wextra -Wpedantic; the flags move into
WARNFLAGS with -Werror so it stays that way. The ARM9 had never seen
-Wextra, which surfaced 203 diagnostics in five categories, all fixed:

- 103 shift-negative-value, all one root cause: libnds's NORMAL_PACK masks
  x and y but shifts z unmasked, undefined for a negative z. general.h now
  redefines it with z masked; the packed field bits are identical.
- 78 unused-parameter, in callback and table-driven signatures that cannot
  change (menu buttons, control handlers, editor entity hooks). Each
  flagged parameter is marked unused in place.
- 16 missing-field-initializers: the entityTypes rows deliberately leave
  model and id for runtime fill, and now say so with a designated
  .modelName initializer; io.c's blankHeader relies on static storage
  being zero instead of spelling out five of its fields.
- 3 type-limits: setBlock checked its u8 coordinates for <0, always false.
  A negative int wraps to >=256 and the size checks catch it.
- 3 sign-compare: getCurrentPortalColor's dist becomes int32 to match z,
  a bound check in font.c gets a u32 cast, and the RLE decoder's loop
  counter becomes uint32_t.

The pre-existing -Wno-implicit-function-declaration and
-Wno-incompatible-pointer-types stay: lifting those is its own project.

Verified with a clean build of both CPUs, the host test suite, and the
portal pipeline smoke test.

+++ arm9/source/font.c
@@ -168,7 +168,7 @@ void loadFont(font_struct* f, u8 charsize, u8 rendersize)
 		for(i=0;i<512/4;i++)
 		{
 
-            if((i*4+3+j*512)<size)
+            if((u32)(i*4+3+j*512)<size)
             {
 			buffer[i+j*512/4]=(pcx.image.data8[i*4+j*512])|((pcx.image.data8[i*4+1+j*512])<<2)|((pcx.image.data8[i*4+2+j*512])<<4)|((pcx.image.data8[i*4+3+j*512])<<6);
             } else
